+++ apps/www/registry/tui/ui/button.tsx
@@ -4,7 +4,8 @@ import { cva, type VariantProps } from "class-variance-authority";
 import { cn } from "@/lib/utils";
 
 
-import { icons, IconsType } from "./helper/icons";
+import Icons from "./icons";
+import { IconsType } from "./helper/fontAwesomeIcon";
 
 const buttonVariants = cva(
   "inline-flex items-center justify-center text-sm font-medium shadow-sm ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50",
@@ -67,7 +68,7 @@ const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
       >
         <div className="flex items-center">
           <div className="mr-2">
-          {typeof icon === "string" && icons(icon)}
+          {typeof icon === "string" && <Icons name={icon} />}
           </div>
           {props.children}
         </div>
